AiAgentDataService/requirements.txt
# logfire==3.6.0
# openai==1.63.2
pydantic==2.10.6
pydantic-ai==0.0.24 
python-dotenv==1.0.1
tabulate==0.9.0
# torch==2.6.0
# transformers==4.48.2
asyncio==3.4.3 
aiokafka==0.12.0
